backend/requirements.txt
# ══════════════════════════════════════════════
# AGENTICA — Backend Dependencies (Optimized)
# Python 3.11+ (Updated for 3.13 compatibility)
# ══════════════════════════════════════════════

# ─────────────────────────────────────────────
# CORE FRAMEWORK
# ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# ─────────────────────────────────────────────
# HTTP CLIENT (OpenRouter API)
# ─────────────────────────────────────────────
httpx==0.28.1

# ─────────────────────────────────────────────
# REAL-TIME (WebSockets)
# ─────────────────────────────────────────────
websockets==14.1
python-socketio==5.12.1

# ─────────────────────────────────────────────
# GRAPH DATABASE (Neo4j - GraphRAG)
# ─────────────────────────────────────────────
neo4j==5.27.0

# ─────────────────────────────────────────────
# MEMORY DATABASE (MongoDB)
# ─────────────────────────────────────────────
motor==3.7.0

# ─────────────────────────────────────────────
# LLM / AI (OpenRouter Compatible)
# ─────────────────────────────────────────────
openai==1.59.6
tiktoken==0.8.0

# ─────────────────────────────────────────────
# DATA PROCESSING
# ─────────────────────────────────────────────
numpy==2.2.2
pandas==2.2.3

# ─────────────────────────────────────────────
# ASYNC UTILITIES
# ─────────────────────────────────────────────
aiofiles==24.1.0

# ─────────────────────────────────────────────
# UNIQUE ID GENERATION
# ─────────────────────────────────────────────
uuid6==2024.1.12

# ─────────────────────────────────────────────
# LOGGING
# ─────────────────────────────────────────────
loguru==0.7.3
rich==13.9.4

# ─────────────────────────────────────────────
# SECURITY
# ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─────────────────────────────────────────────
# DEVELOPMENT ONLY (Optional)
# ─────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
